Fix multiple llm_hass_api and auto migrate v1
