+++ custom_components/deepseek_conversation/translations/fr.json
@@ -38,6 +38,9 @@
                     "thinking_enabled": "Activer le raisonnement",
                     "strip_markdown": "Supprimer le Markdown pour les assistants vocaux",
                     "reasoning_effort": "Effort de raisonnement"
+                },
+                "data_description": {
+                    "llm_hass_api": "Aucune sélection = aucun contrôle"
                 }
             }
         }
